Rename BivrostBase to BivrostKit
