+++ Sources/BivrostKit/BivrostKit.swift
@@ -1,0 +1,10 @@
+//
+//  Bivrost.swift
+//  BivrostKit
+//
+//  Created by Luis Reisewitz on 10.10.17.
+//
+
+public struct BivrostKit {
+    
+}
