+++ Sources/BivrostKit/Export/BaseDecoder.swift
@@ -1,6 +1,6 @@
 //
 //  BaseDecoder.swift
-//  BivrostBase
+//  BivrostKit
 //
 //  Created by Luis Reisewitz on 10.10.17.
 //
