+++ Sources/BivrostKit/Export/BaseEncoder.swift
@@ -1,6 +1,6 @@
 //
 //  BaseEncoder.swift
-//  BivrostBase
+//  BivrostKit
 //
 //  Created by Luis Reisewitz on 10.10.17.
 //
