+++ Sources/BivrostKit/Export/Protocols/StaticType.swift
@@ -1,15 +1,15 @@
 //
-//  DynamicType.swift
-//  BivrostBase
+//  StaticType.swift
+//  BivrostKit
 //
 //  Created by Luis Reisewitz on 10.10.17.
 //
 
 /// To be used as a marker protocol as to not always have to add the `isDynamic` var.
-/// Can only be used if the type is always dynamic.
-protocol DynamicType: SolidityCodable {}
-extension DynamicType {
+/// Can only be used if the type is always static.
+protocol StaticType: SolidityCodable {}
+extension StaticType {
     static var isDynamic: Bool {
-        return true
+        return false
     }
 }
