+++ Sources/BivrostKit/Export/Types/Bytes.swift
@@ -1,6 +1,6 @@
 //
 //  Bytes.swift
-//  BivrostBase
+//  BivrostKit
 //
 //  Created by Luis Reisewitz on 27.09.17.
 //  Copyright © 2017 Gnosis. All rights reserved.
@@ -8,6 +8,7 @@
 
 import Foundation
 import BigInt
+import CryptoSwift
 
 extension Solidity {
     public struct Bytes {
