+++ Sources/BivrostKit/Parser/ParsingError.swift
@@ -1,6 +1,6 @@
 //
 //  ParsingError.swift
-//  BivrostBase
+//  BivrostKit
 //
 //  Created by Luis Reisewitz on 10.10.17.
 //
