+++ Tests/BivrostKitTests/Assets.swift
@@ -1,6 +1,6 @@
 //
 //  Assets.swift
-//  BivrostBaseTests
+//  BivrostKitTests
 //
 //  Created by Luis Reisewitz on 27.09.17.
 //  Copyright © 2017 Gnosis. All rights reserved.
