+++ Tests/BivrostKitTests/ParameterParserSpec.swift
@@ -1,6 +1,6 @@
 //
 //  ParameterParserSpec.swift
-//  BivrostBaseTests
+//  BivrostKitTests
 //
 //  Created by Luis Reisewitz on 18.09.17.
 //  Copyright © 2017 Gnosis. All rights reserved.
@@ -8,7 +8,7 @@
 
 import Quick
 import Nimble
-@testable import BivrostBase
+@testable import BivrostKit
 
 class ParameterParserSpec: QuickSpec {
     override func spec() {
