+++ Tests/BivrostKitTests/SolidityTypesSpec.swift
@@ -1,6 +1,6 @@
 //
 //  SolidityTypesSpec.swift
-//  BivrostBaseTests
+//  BivrostKitTests
 //
 //  Created by Luis Reisewitz on 25.09.17.
 //  Copyright © 2017 Gnosis. All rights reserved.
@@ -9,7 +9,7 @@
 
 import Quick
 import Nimble
-@testable import BivrostBase
+@testable import BivrostKit
 import BigInt
 import Foundation
